add error handling

+++ controllers/TodoController.js
@@ -1,7 +1,7 @@
 const { Todo } = require('../models/index')
 
 class TodoController {
-    static create(req, res) {
+    static create(req, res, next) {
         let input = {
             title: req.body.title,
             description: req.body.description,
@@ -13,7 +13,7 @@ class TodoController {
                 res.status(201).json({ msg: 'Todo created.', data: result })
             })
             .catch(err => {
-                res.status(500)
+                next(err)
             })
     }
 
@@ -27,18 +27,22 @@ class TodoController {
             })
     }
 
-    static findOne(req, res) {
+    static findOne(req, res, next) {
         let id = req.params.id
         Todo.findByPk(id)
             .then(result => {
-                res.status(200).json(result)
+                if (result) {
+                    res.status(200).json(result)
+                } else {
+                    next({ name: '404NotFound' })
+                }
             })
             .catch(err => {
-                res.status(500)
+                next(err)
             })
     }
 
-    static update(req, res) {
+    static update(req, res, next) {
         let input = {
             title: req.body.title,
             description: req.body.description,
@@ -46,26 +50,41 @@ class TodoController {
             due_date: req.body.due_date
         }
         let id = req.params.id
-        Todo.update(input, { where: { id }, returning: true })
-            .then(result => {
-                console.log(result)
-                res.status(200).json({ msg: 'Todo updated.', data: result })
-            })
+        Todo.findByPk({ where: { id } }).then(todo => {
+            if (todo) {
+                Todo.update(input, { where: { id }, returning: true })
+                    .then(result => {
+                        res.status(200).json({ msg: 'Todo updated.', data: result })
+                    })
+                    .catch(err => {
+                        res.send(err)
+                    })
+            } else {
+                next({ name: '404NotFound' })
+            }
+        })
 
-            .catch(err => {
-                res.send(err)
-            })
     }
 
-    static delete(req, res) {
+    static delete(req, res, next) {
         let id = req.params.id
-        Todo.destroy({ where: { id } })
+        Todo.findByPk(id)
             .then(result => {
-                res.status(200).json({ msg: 'Todo deleted.' })
-            })
+                if (result) {
+                    let data = result
+                    Todo.destroy({ where: { id } })
+                        .then(result => {
+                            res.status(200).json({ msg: 'Todo deleted.', data })
+                        })
+                        .catch(err => {
+                            next(err)
+                        })
+                } else {
+                    next({ name: '404NotFound' })
+                }
 
-            .catch(err => {
-                res.send(err)
+            }).catch(err => {
+                next(err)
             })
     }
 }
